Fix Team Formation Bugs

+++ lib/cadet/accounts/teams.ex
@@ -10,7 +10,7 @@ defmodule Cadet.Accounts.Teams do
   import Ecto.Query
 
   alias Cadet.Repo
-  alias Cadet.Accounts.{Team, TeamMember, CourseRegistration}
+  alias Cadet.Accounts.{Team, TeamMember, CourseRegistration, Notification}
   alias Cadet.Assessments.{Answer, Assessment, Submission}
 
   @doc """
@@ -155,14 +155,25 @@ defmodule Cadet.Accounts.Teams do
 
   """
   def delete_team(%Team{} = team) do
-    Submission
+    submission = Submission
     |> where(team_id: ^team.id)
-    |> Repo.all()
-    |> Enum.each(fn x ->
-      Answer
-      |> where(submission_id: ^x.id)
+    |> Repo.one()
+
+    if submission do
+      Submission
+      |> where(team_id: ^team.id)
+      |> Repo.all()
+      |> Enum.each(fn x ->
+        Answer
+        |> where(submission_id: ^x.id)
+        |> Repo.delete_all()
+      end)
+
+      Notification
+      |> where(submission_id: ^submission.id)
       |> Repo.delete_all()
-    end)
+    end
+
     team
     |> Repo.delete()
   end
